+++ Source/InventoryChanger/GameItems/Storage.h
@@ -23,7 +23,7 @@ public:
     void addCollectible(EconRarity rarity, WeaponId weaponID, bool isOriginal, std::string_view iconPath);
     void addVanillaSkin(WeaponId weaponID, std::string_view iconPath);
     void addServiceMedal(EconRarity rarity, std::uint32_t year, WeaponId weaponID, std::string_view iconPath);
-    void addTournamentCoin(EconRarity rarity, WeaponId weaponID, std::uint32_t tournamentEventID, std::string_view iconPath);
+    void addTournamentCoin(EconRarity rarity, WeaponId weaponID, std::uint8_t tournamentEventID, std::string_view iconPath);
     void addPaintKit(int id, ItemName name, float wearRemapMin, float wearRemapMax);
     void addGlovesWithLastPaintKit(EconRarity rarity, WeaponId weaponID, std::string_view iconPath);
     void addSkinWithLastPaintKit(EconRarity rarity, WeaponId weaponID, std::string_view iconPath);
@@ -33,8 +33,8 @@ public:
     void addCaseKey(EconRarity rarity, WeaponId weaponID, std::string_view iconPath);
     void addOperationPass(EconRarity rarity, WeaponId weaponID, std::string_view iconPath);
     void addStatTrakSwapTool(EconRarity rarity, WeaponId weaponID, std::string_view iconPath);
-    void addSouvenirToken(EconRarity rarity, WeaponId weaponID, std::uint32_t tournamentEventID, std::string_view iconPath);
-    void addViewerPass(EconRarity rarity, WeaponId weaponID, std::uint32_t tournamentEventID, std::string_view iconPath);
+    void addSouvenirToken(EconRarity rarity, WeaponId weaponID, std::uint8_t tournamentEventID, std::string_view iconPath);
+    void addViewerPass(EconRarity rarity, WeaponId weaponID, std::uint8_t tournamentEventID, std::string_view iconPath);
 
     [[nodiscard]] const auto& getStickerKit(const Item& item) const
     {
@@ -93,10 +93,10 @@ public:
         return static_cast<bool>(collectible.getDataIndex());
     }
 
-    [[nodiscard]] std::uint32_t getTournamentEventID(const Item& item) const noexcept
+    [[nodiscard]] std::uint8_t getTournamentEventID(const Item& item) const noexcept
     {
         assert(item.isSouvenirToken() || item.isViewerPass() || item.isTournamentCoin());
-        return static_cast<std::uint32_t>(item.getDataIndex());
+        return static_cast<std::uint8_t>(item.getDataIndex());
     }
 
     [[nodiscard]] bool hasPaintKit(const Item& item) const noexcept
